% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{aggregates}
\alias{aggregates}
\title{Get aggregate bars for a stock over a time period}
\usage{
aggregates(
  ticker,
  from,
  to,
  timespan = "day",
  multiplier = 1,
  api_key = get_api_key(),
  adjusted = TRUE,
  limit = 50000,
  sort = "asc",
  rate_limit = 5,
  max_reqs = 5
)
}
\arguments{
\item{ticker}{Specify a case-sensitive ticker symbol. For example, "AAPL"
represents Apple Inc.}

\item{from}{The start of the aggregate time window. Either a date with the
format "YYYY-MM-DD" or a millisecond timestamp.}

\item{to}{The end of the aggregate time window. Either a date with the format
"YYYY-MM-DD" or a millisecond timestamp.}

\item{timespan}{The size of the time window.}

\item{multiplier}{The size of the timespan multiplier.}

\item{api_key}{String containing API key for polygon.io.}

\item{adjusted}{Whether or not the results are adjusted for stock splits.}

\item{limit}{Limits the number of base aggregates that are used to create an
aggregates request. Max 50000. See \code{max_reqs} for further information about
queries that exceed \code{limit}.}

\item{sort}{Sort the results by timestamp. "asc" will return results in
ascending order (oldest at the top), "desc" will return results in
descending order (newest at the top).}

\item{rate_limit}{Number of API requests allowed per minute. Default = 5
corresponds to the 5 requests allowed per minute for Basic plan users. All
paid plan users are allowed unlimited requests per minute, equivalent to
rate_limit = \code{Inf}.}

\item{max_reqs}{The number of requests allowed for a single query. If a query
exceeds the maximum size for a single request then \code{query} will iteratively
send requests until either the query is completed or \code{max_reqs} requests
are sent. See \code{limit} to increase the maximum request size.}
}
\value{
A tibble containing price information of \code{ticker} for each of the
requested periods.
}
\description{
Use the aggregates API to get aggregate bars for a particular stock over a
given date range in custom time window sizes.
}
\references{
\insertRef{stocksDocumentation}{polygonR}
}
