% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_api_key}
\alias{get_api_key}
\title{Get the previously set polygon.io api key}
\usage{
get_api_key()
}
\value{
An api key or fails if nothing is set
}
\description{
Get the previously set polygon.io api key
}
\examples{
key <- get_api_key()
}
