% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{open_close}
\alias{open_close}
\title{Get the open and closing prices of an asset}
\usage{
open_close(
  ticker,
  date,
  adjusted = TRUE,
  api_key = get_api_key(),
  rate_limit = 5
)
}
\arguments{
\item{ticker}{A string containing the ticker code for an asset (stock,
option, index, etc.). See \code{\link{ticker_type}} for more details on formatting
ticker codes.}

\item{date}{Either a date with the format "YYYY-MM-DD" or a millisecond
timestamp.}

\item{adjusted}{Whether or not the results are adjusted for stock splits.}

\item{api_key}{String containing API key for polygon.io.}

\item{rate_limit}{Number of API requests allowed per minute. Default = 5
corresponds to the 5 requests allowed per minute for Basic plan users. All
paid plan users are allowed unlimited requests per minute, equivalent to
rate_limit = \code{Inf}.}
}
\value{
A tibble containing the requested data.
}
\description{
Get the open, close and after-hours prices of an asset on a particular date.
}
\references{
\insertRef{stocksDocumentation}{polygonR}
}
