% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonR-package.R, R/senateR-package.R
\docType{package}
\name{polygonR-package}
\alias{polygonR-package}
\title{polygonR: Interact With the Polygon REST API}
\description{
Provide functions for querying the polygon.io REST API. Queries return 'tidy' data.

Provide functions for querying the polygon.io REST API. Queries return 'tidy' data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flynngo/polygonR}
  \item Report bugs at \url{https://github.com/flynngo/polygonR/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/flynngo/polygonR}
  \item Report bugs at \url{https://github.com/flynngo/polygonR/issues}
}

}
\author{
\strong{Maintainer}: Flynn Gewirtz-O'Reilly \email{f.gewirtz-oreilly.1@research.gla.ac.uk}

}
\keyword{internal}
