% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{process_agg}
\alias{process_agg}
\title{Convert polygon.io aggregates query from json to tidy format}
\usage{
process_agg(json)
}
\arguments{
\item{json}{response object from \code{\link[httr2:resp_body_raw]{httr2::resp_body_json}}}
}
\value{
tibble containing information in \code{json}.
}
\description{
Convert the json obtained from an aggregates query of polygon.io into tabular
data.
}
