% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{request_agg}
\alias{request_agg}
\title{Request data from polygon aggregates api}
\usage{
request_agg(
  ticker,
  from,
  to,
  timespan = "day",
  multiplier = 1,
  api_key = get_api_key(),
  adjusted = TRUE,
  limit = 50000,
  sort = "asc"
)
}
\arguments{
\item{ticker}{Specify a case-sensitive ticker symbol. For example, AAPL represents Apple Inc.}

\item{from}{The start of the aggregate time window. Either a date with the format YYYY-MM-DD or a millisecond timestamp.}

\item{to}{The end of the aggregate time window. Either a date with the format YYYY-MM-DD or a millisecond timestamp.}

\item{timespan}{The size of the time window.}

\item{multiplier}{The size of the timespan multiplier.}

\item{api_key}{polygon.io API key}

\item{adjusted}{Whether or not the results are adjusted for splits.}

\item{limit}{Limits the number of base aggregates queried to create the aggregate results. Max 50000.}

\item{sort}{Sort the results by timestamp. "asc" will return results in ascending order (oldest at the top), "desc" will return results in descending order (newest at the top).}
}
\description{
Requests data using the polygon.io Aggregates API.
}
\references{
https://polygon.io/docs/stocks/get_v2_aggs_ticker__stocksticker__range__multiplier___timespan___from___to
}
