% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_api_key}
\alias{set_api_key}
\title{Set polygon.io api key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{The api key as a string}
}
\value{
void
}
\description{
Set polygon.io api key
}
\examples{
set_api_key("my_api_key")
}
