% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{ticker_type}
\alias{ticker_type}
\title{Detect market type of a ticker}
\usage{
ticker_type(ticker)
}
\arguments{
\item{ticker}{A string containing a ticker code.}
}
\value{
One of \code{"stock"}, \code{"option"}, \code{"fx"}, \code{"index"} and \code{"crypto"}.
}
\description{
Classify a \code{ticker} as either stock, option, forex, index or crypto based on
it's format.
}
\details{
Only the ticker prefix is considered when detecting the market type. This
function does not check that \code{ticker} corresponds to a valid asset, see
Section: \emph{Ticker formats} for more details on correctly formatting ticker
codes.
}
\section{Ticker formats}{
Tickers of each type are formatted as follows.
\itemize{
\item \strong{Stock} tickers are simply formatted as standard ticker codes. E.g. \code{"AAPL"}
for Apple.
\item \strong{Option} tickers follow the general format \code{"O:AAPL211119C00085000"}, where
\itemize{
\item O: indicates that this is an option ticker,
\item AAPL is the ticker for the underlying stock,
\item 211119 is the expiration date in YYMMDD format,
\item C or P indicates whether the option is a call or put option, and
\item 00085000 is the strike price to three decimal places, in this case $85.
}
\item \strong{Forex} price tickers follow the general format \code{"C:GBPUSD"}, where
\itemize{
\item C: indicates that this is a forex ticker,
\item GBP is the currency we are exchanging from, in this case GB
pounds, and
\item USD is the currency we are exchanging to, in this case US dollars.
}
\item \strong{Index} tickers follow the general format \code{"I:NDX"}, where:
\itemize{
\item I: indicates that this is an index ticker,
\item NDX is the the index ticker code, in this case the NASDAQ 100.
}
\item \strong{Crypto} price tickers follow the general format \code{"X:BTCUSD"}, where
\itemize{
\item X: indicates that this is crypto ticker,
\item BTC is the currency we are exchanging from, in this case bitcoin, and
\item USD is the currency we are exchanging to.
}
}
}

\examples{
ticker_type("AAL")
}
\seealso{
\href{https://polygon.io/blog/how-to-read-a-stock-options-ticker}{\emph{polygon.io}: How to Read A Stock Options Ticker}
}
