% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{tidy_aggregates}
\alias{tidy_aggregates}
\title{Convert polygon.io aggregates query to tidy format}
\usage{
tidy_aggregates(resp)
}
\arguments{
\item{resp}{query response (see \code{\link[httr2:response]{httr2::response}})}
}
\value{
tibble containing data in \code{resp}.
}
\description{
Convert the response from an aggregates query of polygon.io into tabular
data.
}
