% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{tidy_results}
\alias{tidy_results}
\title{Format results attribute as tidy data}
\usage{
tidy_results(results)
}
\arguments{
\item{results}{results attribute from query (see \code{\link[=process_agg]{process_agg()}}).}
}
\value{
results attribute as a tibble
}
\description{
Convert the results attribute into tabular data.
}
