% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{grouped_daily}
\alias{grouped_daily}
\alias{stocks_daily}
\alias{crypto_daily}
\alias{fx_daily}
\title{Get the daily open, high, low, and close for a market}
\usage{
grouped_daily(
  date,
  market = "stocks",
  include_otc = FALSE,
  api_key = get_api_key(),
  adjusted = TRUE,
  rate_limit = 5
)

stocks_daily(
  date,
  include_otc = FALSE,
  api_key = get_api_key(),
  adjusted = TRUE,
  rate_limit = 5
)

crypto_daily(date, api_key = get_api_key(), adjusted = TRUE, rate_limit = 5)

fx_daily(date, api_key = get_api_key(), adjusted = TRUE, rate_limit = 5)
}
\arguments{
\item{date}{Either a date with the format "YYYY-MM-DD" or a millisecond
timestamp.}

\item{market}{The market to show, possible values are "stocks", "fx", "forex", "crypto".}

\item{include_otc}{Include OTC securities (default = \code{FALSE}).}

\item{api_key}{String containing API key for polygon.io.}

\item{adjusted}{Whether or not the results are adjusted for stock splits.}

\item{rate_limit}{Number of API requests allowed per minute. Default = 5
corresponds to the 5 requests allowed per minute for Basic plan users. All
paid plan users are allowed unlimited requests per minute, equivalent to
rate_limit = \code{Inf}.}
}
\value{
A tibble containing the requested data.
}
\description{
Get the daily open, high, low, and close (OHLC) for an entire market on a
particular date.
}
\references{
\insertRef{stocksDocumentation}{polygonR}
}
