% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senateR-package.R
\docType{package}
\name{polygonR-package}
\alias{polygonR}
\alias{polygonR-package}
\title{polygonR: Interact With the Polygon REST API}
\description{
Provide functions for querying the polygon.io REST API. Queries return 'tidy' data.
}
\author{
\strong{Maintainer}: Flynn Gewirtz-O'Reilly \email{f.gewirtz-oreilly.1@research.gla.ac.uk}

}
\keyword{internal}
