% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{prev_close}
\alias{prev_close}
\title{Get the previous day's open, high, low, and close (OHLC) for a stock.}
\usage{
prev_close(ticker, adjusted = TRUE, api_key = get_api_key(), rate_limit = 5)
}
\arguments{
\item{ticker}{Specify a case-sensitive ticker symbol. For example, "AAPL"
represents Apple Inc.}

\item{adjusted}{Whether or not the results are adjusted for stock splits.}

\item{api_key}{String containing API key for polygon.io.}

\item{rate_limit}{Number of API requests allowed per minute. Default = 5
corresponds to the 5 requests allowed per minute for Basic plan users. All
paid plan users are allowed unlimited requests per minute, equivalent to
rate_limit = \code{Inf}.}
}
\value{
A tibble containing the requested data.
}
\description{
Get the previous day's open, high, low, and close (OHLC) for a stock.
}
\examples{
prev_close("AAPL")
}
\references{
\insertRef{stocksDocumentation}{polygonR}.
}
