% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API key as environment variable}
\usage{
set_api_key(key = NULL)
}
\arguments{
\item{key}{Default = \code{NULL} will prompt the user to enter their API key and
is recommended. Alternatively users can pass their API key as a string,
although this will print the key in the console, which is less secure.}
}
\description{
Set API key as environment variable
}
